correct paths for moved interfaces and add aria-label to facility icons
