+++ src/components/venue/elements/facilities/Breakfast.tsx
@@ -3,14 +3,14 @@ import { restaurantIcon } from "../../../../assets/icon";
 function BreakfastFacility() {
   return (
     <div className="d-flex me-2 me-md-5 mt-1">
-      <img className="facility-icon me-2" src={restaurantIcon} alt="" />
+      <img className="facility-icon me-2" src={restaurantIcon} aria-label="breakfast available" />
       <p className="fs-0-75rem-to-1rem">Restaurant</p>
     </div>
   );
 }
 
 function BreakfastFacilityNoText() {
-  return <img className="facility-icon me-2" src={restaurantIcon} alt="" />;
+  return <img className="facility-icon me-2" src={restaurantIcon} aria-label="breakfast available" />;
 }
 
 export { BreakfastFacility, BreakfastFacilityNoText };
